<add camera test>

+++ fury_server/server_tcp.cpp
@@ -17,15 +17,17 @@
 #define PORT DEFAULT_PORT
 #define MAX_BUFFER BUFSIZE//1024
 
+#define EXEC_CAMERA_CMD ". /home/android/pi/github/mjpg-streamer/exec &"
 
 int server_sockfd = 0;
+int camera_work_flag = 0;
 
-int exec_system_call(char *cmd) 
-{
+int exec_system_call(char *cmd) {
     FILE *pp = NULL;
 
-    printf("exec_system_call: %s\n",cmd);
+    printf("system call: %s\n",cmd);
     if((pp = popen(cmd, "r")) == NULL) {
+        printf("exec :%s fail!\n", cmd);
         return -1;
     }
 
@@ -33,8 +35,7 @@ int exec_system_call(char *cmd)
     return 0;
 }
 
-int get_pid_by_proc_name(char *proc) 
-{
+int get_pid_by_proc_name(char *proc) {
     char tmp[50] = {0};
     char cmd[50] = {0};
     FILE *pp;
@@ -54,35 +55,48 @@ int get_pid_by_proc_name(char *proc)
     return atoi(tmp);
 }
 
-int send_signal_to_proc(int sig, char *proc) 
-{
+int send_signal_to_proc(int sig, char *proc) {
     int pid = get_pid_by_proc_name(proc);
-    return kill(pid,sig);
+    if(pid > 0) {
+        printf("kill proc:%s\n", proc);
+        return kill(pid,sig);
+    }
+    printf("kill fail!\n");
+    return -1;
 }
 
-void signalHandler(int sig) 
-{
+void signalHandler(int sig) {
     if(server_sockfd > 0) {
+        printf("Opps! release socket resource! \n");
         close(server_sockfd);
         server_sockfd = 0;
-        printf("----close sock fd----\n");
+		stop_camera();
     }
     exit(0);
 }
 
-static void *work_thread(void *args) 
-{
-    //exec_system_call("/home/pi/mjpg.sh");
+static void *work_thread(void *args) {
+    if(exec_system_call(EXEC_CAMERA_CMD) == 0) {
+        camera_work_flag = 1;
+    }
     return NULL;
 }
 
 void start_camera(void) {
-    pthread_t thd = 0;
-    pthread_create(&thd, NULL, work_thread, NULL);
-    pthread_join(thd, NULL); 
+    if(camera_work_flag == 0) {
+        pthread_t thd = 0;
+        pthread_create(&thd, NULL, work_thread, NULL);
+        pthread_join(thd, NULL); 
+    }
 }
 
-//int main(int argc,char *argv[])
+void stop_camera(void) {
+    if(camera_work_flag == 1) {
+        send_signal_to_proc(SIGINT, "mjpg_streamer");
+        camera_work_flag = 0;
+    }
+}
+
 int server_socket_init(void)
 {
     struct sockaddr_in server_addr;
@@ -90,19 +104,20 @@ int server_socket_init(void)
     server_addr.sin_addr.s_addr = htonl(INADDR_ANY); 
     server_addr.sin_port = htons(PORT);
     int client_sockfd =0;
-	int on = 1; 
-	int need_connect = 1;
-
+    
     int size;
     char buffer[MAX_BUFFER + 1];
     struct sockaddr_in client_addr;
     socklen_t len = sizeof(client_addr);
-
-    //reigster ctrl+c /ctrl+z signal.
+    char reply[] = "ok";
+    
+    //reigster ctrl+c ,ctrl+z signal.
     signal(SIGINT, signalHandler); 
     signal(SIGTSTP, signalHandler); 
-
+    
     server_sockfd = socket(AF_INET, SOCK_STREAM, 0);
+
+    int on = 1; 
     if((setsockopt(server_sockfd,SOL_SOCKET,SO_REUSEADDR,&on,sizeof(on))) < 0) {  
         perror("setsockopt failed");  
         exit(EXIT_FAILURE);  
@@ -117,16 +132,12 @@ int server_socket_init(void)
         mini_car_gpio_init();
     #endif
 
-    static char reply[] = "ok";
-
     while(1) 
     {
-        if(need_connect) {
+        if(!client_sockfd) {
             printf("waiting connection...\n");
             client_sockfd = accept(server_sockfd, (struct sockaddr *)&client_addr, &len);
-            need_connect = 0;
 			printf("connect client send 'ok' to client!\n");
-			
             write(client_sockfd, reply, 2);
 			start_camera();
         }
@@ -146,8 +157,7 @@ int server_socket_init(void)
             printf("disconnect client !\n");
             close(client_sockfd);
             client_sockfd = 0;
-            need_connect = 1;
-            //send_signal_to_proc(SIGINT, "mjpg_streamer");
+            stop_camera();
         }
     }
